Add cookie-based sessions (#3)
* Add cookie-based sessions

Uses the `actix-session` crate to create rudimentary cookie-based sessions. Eventually, we will want to store the session data in persistent, server-side storage such as Redis.

There is also a "functional" login page, which includes a request-response transaction with the server. However, the server does not check that the credentials are valid (in fact, credentials are not even stored yet).

* Use dotenv crate to fetch environment variables

The `dotenv` crate provides a function that returns the value of an environment variable at runtime. Using this function, we no longer need to call `dotenv().ok()` in .