Add command-line arguments and hierarchical configuration files (#6)
* Add command-line arguments and hierarchical configuration files

Adds the `clap` crate, which makes it easy to parse command-line arguments that are passed to the server binary. Also adds the `config` crate, which we use to read non-secret configuration files from a hierarchy of files and environment variables. The precedence of configuration sources is listed in `server/src/config.rs`.

* Don't set RUST_LOG in npm start

Log verbosity is set through command-line arguments, and defaults to INFO. So, we don't need to set it via an environment variable.